requirements.txt
aiogram==3.7.0
aiohttp==3.9.5
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
pycryptodome==3.20.0
python-dotenv==1.0.1
apscheduler==3.10.4
beautifulsoup4==4.12.3
jinja2==3.1.4
python-multipart==0.0.9
